Merged in feature/KAN-56-updated-to-live-incidents-system (pull request #841)
KAN-56 Updates to live incidents system

* Improves add incident form draft handling.

Adds functionality to save and continue drafts on the add incident form.

This change introduces a mechanism to hide sections of the form after a draft is saved, and allows users to "continue to report" from the saved draft. This involves reordering form fields and trimming irrelevant sections when a draft exists. This improves the user experience by guiding the user through the reporting process in a more intuitive way.

Updates upsert incident to ensure date and description fields have defaults.

* KAN-55 Refactors authentication token handling

Improves token management using React Context.

The changes introduce a dedicated TokenContext for managing authentication tokens, including fetching, storing, and refreshing. It enhances the ApolloProvider by simplifying token injection into GraphQL requests and subscriptions. It also addresses an issue where the UserProvider was fetching data multiple times by skipping the query if a token is not available.