Merged in bugfix/cache-build-changes
Improves the handling of authentication errors by redirecting the user to the sign-in page when a 401 error is encountered. This ensures a better user experience when their session has expired or they lack the necessary permissions.

Implements cache busting by adding a timestamp to the generated asset filenames. This ensures that users always receive the latest version of the application.