Merged in bugfix/various-jd-fixes (pull request #868)
Bugfix/various jd fixes

* Updates dashboard editor layout constraints.

Updates the dashboard editor to allow widgets to span the entire width of the grid, and ensures a minimum height for the editor.

This change provides a more flexible and user-friendly dashboard layout.

* Increases width of "New Activity" modal

The "New Activity" modal's width is increased to 800 pixels.
This provides more space for the content within the modal,
potentially improving the user experience.

* Enhances caching and build information

Improves caching strategy for HTML files by adding `no-cache`, `no-store`, `Pragma`, and `Expires` headers, ensuring the browser always requests the latest version.

Adds build date to the application for better version tracking and debugging.