Merged in feature/KAN-74-frontend-changes (pull request #892)
KAN-74 Frontend changes

* KAN-60 Adds currency support to stock items

Adds currency support to stock items by fetching the business currency and passing it to the StockItemSelect component.

This ensures that stock item prices are displayed in the correct currency. The currency is fetched from the business object and defaults to the scheme currency if not available.

* KAN-123 checklist template can have roles assigned

* Adds checklist export functionality.

Adds the ability to export checklists to a CSV zip file.

This includes:
- A new view for configuring the export (date range, businesses, users)
- GraphQL mutations and queries for generating and retrieving the data
- Route configuration.

The data export feature allows users to extract checklist data for external analysis and reporting.

* Adds currency support to goods forms

Adds currency support to AddGoods and EditGoods forms, displaying the appropriate currency symbol based on the business or scheme settings.

This ensures that the correct currency is displayed when adding or editing goods, improving usability and data accuracy.

* KAN-51 Imports MySafety data via document upload

Removes the MySafetyCSV component and related incident data structures.

Implements document upload functionality for importing MySafety data via a CSV file.
The CSV is uploaded to a storage and then processed server-side.

Adds a redirect to the homepage after a new scheme is selected.