Merged in feature/init-recycle (pull request #924)
Adds pagination to recycled items and fixes types

* Adds pagination to recycled items and fixes types

Implements pagination for recycled items to improve performance and user experience when dealing with large datasets.

Adds a query to fetch the count of recycled items for pagination purposes.

Updates recycled item types to allow null values for `systemTask`, `incident`, and `offender`.