Merged in feature/stock-improvements (pull request #946)
Refactors StockItemSelect to use stockItemsSearch

* Refactors StockItemSelect to use stockItemsSearch

Updates StockItemSelect to use the new stockItemsSearch query, replacing stockItemsRelay. This change improves the search functionality and simplifies the component's data fetching logic.
The StockItemSearch component is marked as deprecated.

Also fixes an issue in the RecycleBin view where a business name could be undefined.