Merged in feature/export-incident (pull request #964)
Adds select all businesses functionality

* Adds select all businesses functionality

Implements a "Select All Businesses" feature in the BusinessesSelect component and Export Incidents view.

This enhancement provides a more convenient way for users to select all businesses at once, improving the user experience in scenarios where exporting data for all businesses is required.
It introduces an `allSelected` prop to the BusinessesSelect to indicate the state of all businesses being selected or not.