Adds Activity Users Select component
Implements a new ActivityUsersSelect component with related GraphQL queries and UI.

This component allows selecting users assigned to activities, fetching user data with pagination and search capabilities. Also includes the addition of @importok libraries to package.json

The change also updates the Todo view and edit components to include the "Alert ID" (reference) and allows todos to be assigned to groups.

Fixes: #3631