Merged in feature/intel-one-change (pull request #970)
Refactors IntelOne import to use URL based upload

* Refactors IntelOne import to use URL based upload

Updates the IntelOne import process to upload the CSV data to a URL instead of directly sending the data.

This improves scalability and reduces the size of the request payload.

A utility function is added to handle saving JSON data to a URL.

* Adds checklist-based workflow triggering

Enables workflows to be triggered based on checklist completion, checklist score, or the use of a specific checklist template.

Removes unnecessary console log.

* Adds 'useCreatedBy' option for workflows

Adds a switch to include the user who created or completed the triggering model in workflows when not scheduled.

This allows for more dynamic workflows that can leverage user context.