Merged in bugfix/activit (pull request #1070)
Adds task time tracking functionality.

* Adds task time tracking functionality.

Enables time tracking for tasks based on scheme configuration.
This commit introduces a `taskTimeTracking` property in the scheme configuration,
which controls the visibility of the time tracking section in the Add/Edit Todo forms.

The time tracking section is now conditionally rendered based on the value of
`taskTimeTracking`. This allows administrators to enable or disable time tracking
on a per-scheme basis.

Also, this commit includes a fix to the LinkProfile component, changing the way
role permissions are checked to use a custom hook, solving the issue of
deprecated functions.
Finally, some new fields are added to the `UpdateTodoDetails` mutation.
* Fix build check